projects/nanobot-memory-improvement/pyproject.toml
[project]
name = "nanobot-memory"
version = "0.1.0"
description = "Multi-scope memory extension for nanobot"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "nanobot-contributors" }
]
keywords = ["nanobot", "ai", "assistant", "memory", "discord"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = []

[project.optional-dependencies]
discord = ["discord.py>=2.0"]
test = ["pytest>=7.0", "pytest-asyncio>=0.21"]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21", "black", "ruff"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
